updated console output and readme with docs/imgs

+++ heic_image_converter/Converter.py
@@ -55,6 +55,7 @@ class Converter(object):
 		# check src dir
 		dir_src_try = cmd_args[0].strip()
 		if os.path.isdir(dir_src_try):
+			print(f"{PRE}{SET}source directory found: '{dir_src_try}'")
 			self.dir_src = dir_src_try
 		else:
 			print(f"{PRE}{SET}there was an issue finding the image source directory: '{dir_src_try}'")
@@ -77,10 +78,10 @@ class Converter(object):
 
 					if flag == "--dest":
 						if not os.path.isdir(flag_value):
-							print(f"{PRE}{SET}did not find the '--dest' directory. creating '{flag_value}'...")
+							print(f"{PRE}{SET}did not find the '--dest' directory: '{flag_value}'")
 							try:
+								print(f"{PRE}{SET}creating it...")
 								os.makedirs(flag_value)
-								print(f"{PRE}{SET}created '{flag_value}'")
 							except Exception as e:
 								print(e)
 								print(f"{PRE}{SET}there was an issue creating the '--dest' directory.")
@@ -99,12 +100,13 @@ class Converter(object):
 					print(f"{PRE}{SET}'{flag}' flag set with '{flag_value}'.")
 		
 
-	def convert_files(self, image_to_ext="PNG"):
+	def convert_files(self, image_to_ext=".png"):
 		"""
 		Iterates through all `.heic` files in the source image directory, converting them according to set flags...
 		"""
 		dir_bytes = os.fsencode(self.dir_src)
-		for file_bytes in os.listdir(dir_bytes):
+		for i, file_bytes in enumerate(os.listdir(dir_bytes)):
+			n = i+1
 			# obtain strings for path and filename
 			filename = os.fsdecode(file_bytes)
 			filepath = os.path.join(self.dir_src, filename)
@@ -122,14 +124,14 @@ class Converter(object):
 			
 			# if file already exists in destination filetype, skip
 			if os.path.isfile(new_filepath):
-				print(f"{PRE}{CONV}skipping - file already exists: '{filepath}'")
+				print(f"{PRE}{CONV}[{n}] - skipping - file already exists: '{new_filename}'")
 				self.complete_conversions.append(filepath)
 				continue
 
 			# file is desired source format, like '.heic', convert to desired format
 			if filename.lower().endswith(self.image_from_ext):
 				try:
-					print(f"{PRE}{CONV}source: '{filepath}'")
+					print(f"{PRE}{CONV}[{n}] - source: '{filename}'")
 
 					# convert file and save
 					heif_file = pillow_heif.read_heif(filepath)
@@ -138,27 +140,23 @@ class Converter(object):
 					image.save(new_filepath, format(saveAs.lower()))
 					
 					# success
-					print(f"{PRE}{CONV}new image: '{new_filepath}'")
+					print(f"{PRE}{CONV}[{n}] - new image: '{new_filename}'")
 					self.complete_conversions.append(new_filepath)
 
 					# if success, and -del flag, remove original
-					print(f"{PRE}{CONV}'-del' flag has been set. removing '{filepath}'...")
-					os.remove(filepath)
+					if self.flags["-del"][0]:
+						print(f"{PRE}{CONV}[{n}] - '-del' flag has been set. removing '{filename}'...")
+						os.remove(filepath)
 				
 				# there was an issue converting this file, note and continue
 				except Exception as e:
-					print(f"{PRE}{CONV}error converting: '{filepath}'")
+					print(f"{PRE}{CONV}[{n}] - error converting: '{filename}'")
+					print(e)
 					self.incomplete_conversions.append(filepath)
-
-
-	def report_to_console(self):
-		"""Displays the completed and incomplete file conversions to the console."""
+		
+		# show final successes
 		lc = len(self.complete_conversions)
 		li = len(self.incomplete_conversions)
-
-		print(f"{PRE}{REP}Reporting file conversions...")
-		for i, inc in enumerate(self.incomplete_conversions):
-			print(f"{PRE}{REP}[{i+1}/{li} INCOMPLETE] - {inc}")
-		for i, com in enumerate(self.complete_conversions):
-			print(f"{PRE}{REP}[{i+1}/{lc} COMPLETE] - {com}")
-		print(f"{PRE}{REP}report done. {lc}/{(lc + li)} complete.")
+		print(f"{PRE}{REP}image source directory: '{self.dir_src}'")
+		print(f"{PRE}{REP}image save destination: '{save_dir}'")
+		print(f"{PRE}{REP}converted {lc}/{(lc + li)} files.")
